django reports

+++ customadmin/models.py
@@ -62,7 +62,7 @@ class FarmRegister(models.Model):
     category_id = models.ForeignKey(Category, on_delete=models.CASCADE, null=True)
     unit_cost = models.IntegerField()
     unit_acre = models.FloatField()
-    total_cost = models.IntegerField()
+    total_cost = models.DecimalField(max_digits=10, decimal_places=2)
     quantity = models.CharField(max_length=50)
     description = models.TextField(null=True, blank=True)
     date_created = models.DateTimeField(auto_now_add=True)
